templates/requirements.txt
aiofiles==24.1.0
fastapi==0.115.12
Jinja2==3.1.6
google-generativeai==0.8.5
python-dotenv==1.1.0
python-multipart==0.0.20
uvicorn==0.34.3
websockets==15.0.1